pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "python-binance>=1.0.28",
    "scikit-learn>=1.6.1",
    "streamlit>=1.43.2",
]
